==> 02_Data_analysis_and_visualization/13_slingshot_analysis.Rmd <==
---
title: "15_slingshot analysis"
output: html_document
date: "2024-05-10"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


```{r}
# load libraries
library(tidyverse)
library(Seurat)
library(condiments)
library(slingshot)
library(destiny)
library(dittoSeq)
library(dplyr)
library(tidyr)
library(missMDA)
library(dismo)
library(RColorBrewer)
library(gridExtra)
library(tradeSeq)
library(ggplot2)
library(cowplot)
```

## 1. Load seurat
```{r}
#load seurat 
dat <- readRDS ("03_dat_seuratObj.rds")
```

# 2. Slingshot
```{r}
# convert dat from a Seurat object to SCE object
dat_sce <- as.SingleCellExperiment(dat, assay = "RNA")

# get UMAP coordinates from dat_sce
dat_sce_umap <- reducedDims(dat_sce)$UMAP %>% as.matrix()

# run Slingshot
dat_slingshot_res0.4 <- slingshot::slingshot(dat_sce_umap, colData(dat_sce)$mrna_cca_snn_res.0.4, start.clus = "0")

# visualize each lineage goes through which clusters
slingLineages(dat_slingshot_res0.4)

# # save
# saveRDS(11_dat_slingshot_res0.4.rds")
```

# 3. Topology test
Topology test is to test whether or not slingshot should be fitted independently for different conditions or not.
```{r}
topology_slingshot_res0.4 <- condiments::topologyTest(sds = dat_slingshot_res0.4, conditions = colData(dat_sce)$age_group)

# saveRDS(12_topology_slingshot_res0.4.RDS")

knitr::kable(topology_slingshot_res0.4)
```
The p-value is 0, very significant. This means a lineage can be fitter independently for each condition (each age group).

```{r}
# draw individual curves
dat_slingshot_res0.4_curve <- slingshot_conditions(dat_slingshot_res0.4,
                                                colData(dat_sce)$age_group, 
                                                approx_points = FALSE,
                                                extend = "n", 
                                                reweight = FALSE, 
                                                reassign = FALSE)

dat_slingshot_res0.4_curve$condition_id <- names(dat_slingshot_res0.4_curve)

dat_slingshot_res0.4_curve$mapping <- matrix(rep(1:4, each = 3), nrow = 4, ncol = 3, byrow = TRUE)

# merge trajectories
dat_slingshot_res0.4_curve2 <- do.call(merge_sds, dat_slingshot_res0.4_curve)

saveRDS(dat_slingshot_res0.4_curve,"13_dat_slingshot_res0.4_curve_individual.rds")
saveRDS(dat_slingshot_res0.4_curve2,"13_dat_slingshot_res0.4_curve_merged.rds")
```

```{r}
# tidy up data to plot trajectories
dat_sce_umap <- dat_sce_umap %>% 
  as.data.frame() %>% 
  rownames_to_column("cell_id")

metaDat <- metaDat %>% rownames_to_column("cell_id")

df <- full_join(dat_sce_umap, metaDat ,by="cell_id")

df <- full_join(
  df %>% dplyr::select(cell_id, UMAP_1, UMAP_2, mrna_cca_snn_res.0.4, age_group),
  slingPseudotime(dat_slingshot_res0.4_curve2) %>% 
    as.data.frame() %>%
    mutate(cell_id = rownames(.))) %>% 
  pivot_longer(starts_with("Lineage"), names_to = "Curve", values_to = "pst") %>% 
  mutate(age_group = factor(age_group, levels = c("young", "midage", "old")))
```

```{r}
## Plot trajectories
pcurves <- ggplot(df, aes(x = UMAP_1, y = UMAP_2, col = age_group)) +
  geom_point(size = .7, alpha = .1) +
  scale_color_brewer(palette = "Set1")


for (age_group in unique(dat_sce$age_group)) {
  
  sds_cond <- dat_slingshot_res0.4_curve[[age_group]]
  
  for (i in 1:4) {
    pcurves <- pcurves +  
      geom_path(data = slingCurves(sds_cond)[[i]]$s[slingCurves(sds_cond)[[i]]$ord, ] %>%
                  as.data.frame() %>%
                  mutate(age_group = age_group), 
                size = 1.5)   
  }
}

pcurves <- pcurves + theme_classic()

ggsave("trajectories.png", pcurves, device = "png",width = 10, height = 6)
```


# 4. Differential progression

Differential progression plot: change of proportion of cells in each age group along the pseudotime 

```{r}
# perform progression test
#progressionTest <- progressionTest(dat_slingshot_res0.4_curve2, conditions = dat_sce$age_group, lineages = TRUE, global = TRUE)

# plot differential progression
ggplot(df, aes(x = pst)) +
  geom_density(alpha = .4, aes(fill = age_group), col = "transparent") +
  geom_density(aes(col = age_group), fill = "transparent", size = 1.5) +
  guides(col = "none") +
  scale_fill_brewer(palette = "Set1") +
  scale_color_brewer(palette = "Set1") +
  labs(x = "Pseudotime", fill = "Type") +
  facet_wrap(~ Curve, scales = "free_x")

#progressionTest_plot<-

#ggsave("08_differential_progression_res0.4.png",pmature,device = "png",width = 9, height = 6)
```

# 5. Differential fate selection

Each cell has a weight and a pseudotime value associated with it. With the weight, we can test if cells will differentiate similarly or differentaly between age grops.

A weight around 0 represents a cell that is mostly assigned to another lineage
A weight around 0.5 represents a cell that is equally assigned to both lineages, i.e. before the
bifurcation.
A weight around 1 represent a cell that is mostly assigned to this lineage.

```{r}
# calculate the weights
differentiationTest <- fateSelectionTest(dat_slingshot_res0.4_curve2, conditions = dat_sce$age_group, pairwise = TRUE)

weights <- condiments:::.sling_reassign(dat_slingshot_res0.4_curve2) %>% 
  as.data.frame() %>% 
  mutate(cell_id = rownames(.)) %>% 
  dplyr::rename("Lineage1" = V1, "Lineage2" = V2, "Lineage3" = V3, "Lineage4" = V4) %>% 
  pivot_longer(starts_with("Lineage"), names_to = "Curve", values_to = "weights") %>% 
  left_join(metaDat, by = "cell_id")

# calculate the average of mean in each age group in each lineage
df_w <- weights %>%
  group_by(cell_id) %>%
  mutate(weights = weights / sum(weights)) %>%
  ungroup() %>%
  group_by(age_group, Curve) %>%
  summarise(weights = mean(weights), .groups = NULL)

weights_mean_plot <- ggplot(df_w, aes(x = Curve, fill = age_group, y = weights)) +
  geom_col(position = "dodge") +
  scale_fill_brewer(palette = "Accent") +
  theme(legend.position = c(.7, .7)) +
  labs(x = "", y = "Mean weight")

# ggsave("result/09_weights_mean.png", weights_mean_plot, device = "png",width = 6, height = 6)
```

```{r}
# Weight probability
weights_prob <- slingCurveWeights(dat_slingshot_res0.4_curve2, as.probs = TRUE) %>% 
  as.data.frame() %>% 
  mutate(cell_id = rownames(.)) %>% 
  pivot_longer(starts_with("Lineage"), names_to = "Curve", values_to = "weights") %>% 
  full_join(df)

# pX<-ggplot(df, aes(x = weights_prob, fill = age_group)) +
weights_prob_plot <- ggplot(weights_prob_plot, aes(x = weights, fill = age_group)) +
  geom_density(alpha = .5) +
  scale_fill_brewer(type = "qual") +
  labs(x = "Curve weight for the all lineages")+
  facet_wrap(~ Curve, scales = "free_x")

# ggsave("result/09_weight_probability.png",weights_prob_plot,device = "png",width = 9, height = 4)
```

```{r}
# Plotting weights on the UMAP using dittoSeq UMAP scatter plots - per lineage

young <- SplitObject(dat, split.by = "age_group")$young
midage <- SplitObject(dat, split.by = "age_group")$midage
old <- SplitObject(dat, split.by = "age_group")$old


df_new_temp <- as.data.frame(colnames(dat)) %>% 
  `colnames<-`("cell_id") %>% 
  mutate(order = 1:length(.$cell_id)) %>% 
  merge(weights_prob, by = "cell_id") %>%
  .[order(.$order),]

df_new_temp_lin1<-df_new_temp[which(df_new_temp$Curve=="Lineage1"),]
df_new_temp_lin2<-df_new_temp[which(df_new_temp$Curve=="Lineage2"),]
df_new_temp_lin3<-df_new_temp[which(df_new_temp$Curve=="Lineage3"),]
df_new_temp_lin4<-df_new_temp[which(df_new_temp$Curve=="Lineage4"),]

df_new_temp_young_lin1<-df_new_temp[which(df_new_temp$age_group=="young" & df_new_temp$Curve=="Lineage1"),]
df_new_temp_young_lin2<-df_new_temp[which(df_new_temp$age_group=="young" & df_new_temp$Curve=="Lineage2"),]
df_new_temp_young_lin3<-df_new_temp[which(df_new_temp$age_group=="young" & df_new_temp$Curve=="Lineage3"),]
df_new_temp_young_lin4<-df_new_temp[which(df_new_temp$age_group=="young" & df_new_temp$Curve=="Lineage4"),]

df_new_temp_midage_lin1<-df_new_temp[which(df_new_temp$age_group=="midage" & df_new_temp$Curve=="Lineage1"),]
df_new_temp_midage_lin2<-df_new_temp[which(df_new_temp$age_group=="midage" & df_new_temp$Curve=="Lineage2"),]
df_new_temp_midage_lin3<-df_new_temp[which(df_new_temp$age_group=="midage" & df_new_temp$Curve=="Lineage4"),]
df_new_temp_midage_lin4<-df_new_temp[which(df_new_temp$age_group=="midage" & df_new_temp$Curve=="Lineage4"),]

df_new_temp_old_lin1<-df_new_temp[which(df_new_temp$age_group=="old" & df_new_temp$Curve=="Lineage1"),]
df_new_temp_old_lin2<-df_new_temp[which(df_new_temp$age_group=="old" & df_new_temp$Curve=="Lineage2"),]
df_new_temp_old_lin3<-df_new_temp[which(df_new_temp$age_group=="old" & df_new_temp$Curve=="Lineage3"),]
df_new_temp_old_lin4<-df_new_temp[which(df_new_temp$age_group=="old" & df_new_temp$Curve=="Lineage4"),]

dat$weight_probs_lin1<-df_new_temp_lin1$weights
dat$weight_probs_lin2<-df_new_temp_lin2$weights
dat$weight_probs_lin3<-df_new_temp_lin3$weights
dat$weight_probs_lin4<-df_new_temp_lin4$weights

young$weight_probs_young_lin1<-df_new_temp_young_lin1$weights
young$weight_probs_young_lin2<-df_new_temp_young_lin2$weights
young$weight_probs_young_lin3<-df_new_temp_young_lin3$weights
young$weight_probs_young_lin4<-df_new_temp_young_lin4$weights

midage$weight_probs_midage_lin1<-df_new_temp_midage_lin1$weights
midage$weight_probs_midage_lin2<-df_new_temp_midage_lin2$weights
midage$weight_probs_midage_lin3<-df_new_temp_midage_lin3$weights
midage$weight_probs_midage_lin4<-df_new_temp_midage_lin4$weights

old$weight_probs_old_lin1<-df_new_temp_old_lin1$weights
old$weight_probs_old_lin2<-df_new_temp_old_lin2$weights
old$weight_probs_old_lin3<-df_new_temp_old_lin3$weights
old$weight_probs_old_lin4<-df_new_temp_old_lin4$weights

# Plotting weights on the UMAP using dittoSeq UMAP scatter plots - per lineage
p1_weights<-dittoDimPlot(dat, "weight_probs_lin1", reduction.use = "umap", min.color = "lightgrey", max.color = "blue", split.by="age_group")
p2_weights<-dittoDimPlot(dat, "weight_probs_lin2", reduction.use = "umap", min.color = "lightgrey", max.color = "blue", split.by="age_group")
p3_weights<-dittoDimPlot(dat, "weight_probs_lin3", reduction.use = "umap", min.color = "lightgrey", max.color = "blue", split.by="age_group")
p4_weights<-dittoDimPlot(dat, "weight_probs_lin4", reduction.use = "umap", min.color = "lightgrey", max.color = "blue", split.by="age_group")

# Plotting weights on the UMAP using dittoSeq UMAP scatter plots - per lineage and per age group
# young
pyoung_lin1<-dittoDimPlot(young, "weight_probs_young_lin1", reduction.use = "umap", min.color = "lightgrey", max.color = "blue", split.by="age_group")
pyoung_lin2<-dittoDimPlot(young, "weight_probs_young_lin2", reduction.use = "umap", min.color = "lightgrey", max.color = "blue", split.by="age_group")
pyoung_lin3<-dittoDimPlot(young, "weight_probs_young_lin3", reduction.use = "umap", min.color = "lightgrey", max.color = "blue", split.by="age_group")
pyoung_lin4<-dittoDimPlot(young, "weight_probs_young_lin4", reduction.use = "umap", min.color = "lightgrey", max.color = "blue", split.by="age_group")

# midage
pmidage_lin1<-dittoDimPlot(midage, "weight_probs_midage_lin1", reduction.use = "umap", min.color = "lightgrey", max.color = "blue", split.by="age_group")
pmidage_lin2<-dittoDimPlot(midage, "weight_probs_midage_lin2", reduction.use = "umap", min.color = "lightgrey", max.color = "blue", split.by="age_group")
pmidage_lin3<-dittoDimPlot(midage, "weight_probs_midage_lin3", reduction.use = "umap", min.color = "lightgrey", max.color = "blue", split.by="age_group")
pmidage_lin4<-dittoDimPlot(midage, "weight_probs_midage_lin4", reduction.use = "umap", min.color = "lightgrey", max.color = "blue", split.by="age_group")

# old #
pold_lin1<-dittoDimPlot(old, "weight_probs_old_lin1", reduction.use = "umap", min.color = "lightgrey", max.color = "blue", split.by="age_group")
pold_lin2<-dittoDimPlot(old, "weight_probs_old_lin2", reduction.use = "umap", min.color = "lightgrey", max.color = "blue", split.by="age_group")
pold_lin3<-dittoDimPlot(old, "weight_probs_old_lin3", reduction.use = "umap", min.color = "lightgrey", max.color = "blue", split.by="age_group")
pold_lin4<-dittoDimPlot(old, "weight_probs_old_lin4", reduction.use = "umap", min.color = "lightgrey", max.color = "blue", split.by="age_group")
```


# 6. TradeSeq

## 1. fitGAM
Before we can fit these smoothed expression profiles, we need to get a sense of how complex the expression patterns are in this dataset. This is translated into selecting a number of knots for the NB-GAMs, where a higher number of knots allows for more complex expression patterns. Here, we pick 8 knots.

We fit the NB-GAMs using 8 knots, based on the pseudotime and cell-level weights estimated by Slingshot. We use the conditions argument to fit separate smoothers for each condition.

```{r}
# TradeSeq calculation is intensive. Move to cluster to calculate.
# dat_sce@int_metadata$slingshot <- dat_slingshot_res0.4_curve2
# saveRDS(dat_sce, "result/14_dat_sce.rds")
```

```{r}
# # example scrip to calculate TradeSeq on cluster
# 
# library(tradeSeq)
# set.seed(3)
# BPPARAM <- BiocParallel::bpparam()
# BPPARAM$workers <- 30
# 
# dat_sce <- fitGAM(dat_sce, conditions = factor(colData(dat_sce)$age_group), nknots = 8)
# 
# saveRDS(dat_sce,"15_tradeseq_res0.4.rds")
```

```{r}
dat_sce <- readRDS("result/15_tradeseq_res0.4.rds")
```

## 2. Discovering progenitor marker genes. 

This function assesses differential expression between the average expression of the start and end points of a lineage.

```{r}
# access differential expression between the average expression of the start and end points of a lineage
startRes <- startVsEndTest(dat_sce, lineages = TRUE)

# lineage 1
startVsEndTest_lineage1 <- startRes %>% 
  dplyr::select(contains("lineage1")) %>% 
  mutate(pva_adj_lineage1 = p.adjust(pvalue_lineage1), "fdr") %>% 
  filter(pva_adj_lineage1 < 0.05) %>% 
  rownames_to_column("gene")

# lineage 2
startVsEndTest_lineage2 <- startRes %>% 
  dplyr::select(contains("lineage2")) %>% 
  mutate(pva_adj_lineage2 = p.adjust(pvalue_lineage2), "fdr") %>% 
  filter(pva_adj_lineage2 < 0.05) %>% 
  rownames_to_column("gene")

# lineage 3
startVsEndTest_lineage3 <- startRes %>% 
  dplyr::select(contains("lineage3")) %>% 
  mutate(pva_adj_lineage3 = p.adjust(pvalue_lineage3), "fdr") %>% 
  filter(pva_adj_lineage3 < 0.05) %>% 
  rownames_to_column("gene")

# lineage 4
startVsEndTest_lineage4 <- startRes %>% 
  dplyr::select(contains("lineage4")) %>% 
  mutate(pva_adj_lineage4 = p.adjust(pvalue_lineage4), "fdr") %>% 
  filter(pva_adj_lineage4 < 0.05) %>% 
  rownames_to_column("gene")

# report gene list
write_csv(startVsEndTest_lineage1, "result/12_progenitor_marker_lin1.csv")
write_csv(startVsEndTest_lineage2, "result/12_progenitor_marker_lin2.csv")
write_csv(startVsEndTest_lineage3, "result/12_progenitor_marker_lin3.csv")
write_csv(startVsEndTest_lineage4, "result/12_progenitor_marker_lin4.csv")
```


Compare specific pseudotime values within a lineage

```{r}
# customRes <- startVsEndTest(dat_sce, pseudotimeValues = c(0.1, 0.8), lineages = TRUE)
```


## 3. Discovering differentiated cell type markers

This function asseses differential expression between the average expression at the end points of lineages.

```{r}
endRes <- diffEndTest(dat_sce, pairwise = TRUE)

# lineage 1 vs lineage 2
endRes_lineage1vs2 <- endRes %>% 
  dplyr::select(contains(c("1vs2", "1_2"))) %>% 
  mutate(pva_adj_1vs2 = p.adjust(pvalue_1vs2), "fdr") %>% 
  filter(pva_adj_1vs2 < 0.05) %>% 
  rownames_to_column("gene")

# lineage 1 vs lineage 3
endRes_lineage1vs3 <- endRes %>% 
  dplyr::select(contains(c("1vs3", "1_3"))) %>% 
  mutate(pva_adj_1vs3 = p.adjust(pvalue_1vs3), "fdr") %>% 
  filter(pva_adj_1vs3 < 0.05) %>% 
  rownames_to_column("gene")

# lineage 2 vs lineage 3
endRes_lineage2vs3 <- endRes %>% 
  dplyr::select(contains(c("2vs3", "2_3"))) %>% 
  mutate(pva_adj_2vs3 = p.adjust(pvalue_2vs3), "fdr") %>% 
  filter(pva_adj_2vs3 < 0.05) %>% 
  rownames_to_column("gene")

# lineage 1 vs lineage 4
endRes_lineage1vs4 <- endRes %>% 
  dplyr::select(contains(c("1vs4", "1_4"))) %>% 
  mutate(pva_adj_1vs4 = p.adjust(pvalue_1vs4), "fdr") %>% 
  filter(pva_adj_1vs4 < 0.05) %>% 
  rownames_to_column("gene")

# lineage 2 vs lineage 4
endRes_lineage2vs4 <- endRes %>% 
  dplyr::select(contains(c("2vs4", "2_4"))) %>% 
  mutate(pva_adj_2vs4 = p.adjust(pvalue_2vs4), "fdr") %>% 
  filter(pva_adj_2vs4 < 0.05) %>% 
  rownames_to_column("gene")

# lineage 3 vs lineage 4
endRes_lineage3vs4 <- endRes %>% 
  dplyr::select(contains(c("3vs4", "3_4"))) %>% 
  mutate(pva_adj_3vs4 = p.adjust(pvalue_3vs4), "fdr") %>% 
  filter(pva_adj_3vs4 < 0.05) %>% 
  rownames_to_column("gene")


# report gene list
write_csv(endRes_lineage1vs2, "result/13_endRes_lineage1vs2.csv")
write_csv(endRes_lineage1vs3, "result/13_endRes_lineage1vs3.csv")
write_csv(endRes_lineage2vs3, "result/13_endRes_lineage2vs3.csv")

write_csv(endRes_lineage1vs4, "result/13_endRes_lineage1vs4.csv")
write_csv(endRes_lineage2vs4, "result/13_endRes_lineage2vs4.csv")
write_csv(endRes_lineage3vs4, "result/13_endRes_lineage3vs4.csv")
```


## 4. Plot
Note: 
15_tradeseq_res0.4.rds contains calculation on each lineage each age group.
15_tradeseq_nocondition_res0.4.rds contains calculation on each lineage, no separation on age group.

### 1. plot in conditions and in lineages
Use the calculation in 15_tradeseq_res0.4.rds

Note: For individual plot, please use the below function.
```{r}
plotSmoothers(dat_sce, assays(dat_sce)$counts, gene = "ABCB1")
```

### 2. plot in lineages
Use the calculation in 15_tradeseq_nocondition_res0.4.rds
```{r}
dat_sce_noCond <- readRDS("result/15_tradeseq_nocondition_res0.4.rds")

# # Note: take a long time to plot. Move this step to HPC.
# pltList_noCond <- list()
# 
# for (i in 1:length(tradeseq_to_plot)) {
#   pltList_noCond[[i]] <- plotSmoothers(dat_sce_noCond, assays(dat_sce_noCond)$counts, gene = tradeseq_to_plot[i]) + ggtitle(tradeseq_to_plot[i])
# }
```

Note: For individual plot, please use the below function.
```{r}
plotSmoothers(dat_sce_noCond, assays(dat_sce_noCond)$counts, gene = "ABCB1")
```

```{r}
Top50_HSC1genes_youngvsold <- c(
  "RBM48", "RPL31", "KLF4", "SNHG20", "HIST1H2BI", "CYBA", "BAG6", "GADD45A", "DUSP6", "DDIT3",
  "CXCR4", "STMN1", "CRHBP", "SMIM24", "FOSL1", "ZNF219", "ZRSR2", "FSCN1", "RGCC", "HRAS",
  "MYCN", "NCF1", "PHF1", "VGLL4", "MAPK1IP1L", "NFKBIB", "ATP1B1", "FGD5", "FAM69B", "MYC",
  "GINS2", "LAPTM4B", "APEX1", "JUN", "SCD", "SPNS2", "PNP", "POLR2G", "HES1", "SLC25A5",
  "CDK6", "HLF", "B2M", "HBA2", "BMI1", "CALN1", "RELA", "ODC1", "DLK1"
)
```


### 3. RUN TOP 50 GENES dif expressed between OLDVSYOUNG for HSC1_Lineage 1 
```{r}
Top50_HSC1genes_youngvsold <- c("RBM48", "RPL31", "KLF4", "SNHG20", "HIST1H2BI", "CYBA", "BAG6", "GADD45A", "DUSP6", "DDIT3",
  "CXCR4", "STMN1", "CRHBP", "SMIM24", "FOSL1", "ZNF219", "ZRSR2", "FSCN1", "RGCC", "HRAS",
  "MYCN", "NCF1", "PHF1", "VGLL4", "MAPK1IP1L", "NFKBIB", "ATP1B1", "FGD5", "FAM69B", "MYC",
  "GINS2", "LAPTM4B", "APEX1", "JUN", "SCD", "SPNS2", "PNP", "POLR2G", "HES1", "SLC25A5",
  "CDK6", "HLF", "B2M", "HBA2", "BMI1", "CALN1", "RELA", "ODC1", "DLK1"
)


# Define colors for curves (assuming the same colors for all genes)
curvesCols <- c( "#0000FFFF","green", "#FF0000FF", "transparent",
                "transparent", "transparent", "transparent", "transparent",
                "transparent", "transparent", "transparent", "transparent")

# Initialize an empty list to store individual plots
plots_list <- list()

# Loop through each gene and create a plot
for (gene in Top50_HSC1genes_youngvsold) {
  # Create the plot for the current gene
  plot <- plotSmoothers(dat_sce, assays(dat_sce)$counts, gene = gene, 
                        curvesCols = curvesCols, border = FALSE) +
    ggplot2::scale_color_manual(values = curvesCols) +
    ggtitle(paste("Lineage 1 - ", gene)) +
    theme(legend.position = "none")  # Remove the legend
  
  # Add the plot to the list
  plots_list[[gene]] <- plot
}

# Create a PDF file to save all the plots
pdf("Lineage1_GMP_TOP50genesHSC1_youngvsold.pdf", width = 8, height = 10)  # Adjust width and height as needed

# Arrange and print all plots into the PDF
#plot_grid(plotlist = plots_list, nrow = 5, ncol = 3)  # Adjust the number of columns as needed

marrangeGrob(grobs=plots_list, nrow=5, ncol=3)
# Close the PDF device
dev.off()
```

###TOP50GENES LINEAGE2
```{r}
# Define colors for curves (assuming the same colors for all genes)
curvesCols <- c( "transparent", "transparent", "transparent","#0000FFFF","green", "#FF0000FF", 
                 "transparent", "transparent", "transparent", "transparent", "transparent",
                "transparent")

# Initialize an empty list to store individual plots
plots_list_lin2 <- list()

# Loop through each gene and create a plot
for (gene in Top50_HSC1genes_youngvsold) {
  # Create the plot for the current gene
  plot_lin2 <- plotSmoothers(dat_sce, assays(dat_sce)$counts, gene = gene, 
                        curvesCols = curvesCols, border = FALSE) +
    ggplot2::scale_color_manual(values = curvesCols) +
    ggtitle(paste("Lineage 2 - ", gene)) +
    theme(legend.position = "none")  # Remove the legend
  
  # Add the plot to the list
  plots_list_lin2[[gene]] <- plot_lin2
}

# Create a PDF file to save all the plots
pdf("Lineage2_MDP_TOP50genesHSC1_youngvsold.pdf", width = 8, height = 10)  # Adjust width and height as needed

# Arrange and print all plots into the PDF
#plot_grid(plotlist = plots_list, nrow = 5, ncol = 3)  # Adjust the number of columns as needed

marrangeGrob(grobs=plots_list_lin2, nrow=5, ncol=3)
# Close the PDF device
dev.off()
```

###TOP50GENES LINEAGE3
```{r}
# Define colors for curves (assuming the same colors for all genes)
curvesCols <- c( "transparent", "transparent", "transparent","transparent", "transparent", "transparent", "#0000FFFF","green", "#FF0000FF", "transparent", "transparent", "transparent")

# Initialize an empty list to store individual plots
plots_list_lin3 <- list()

# Loop through each gene and create a plot
for (gene in Top50_HSC1genes_youngvsold) {
  # Create the plot for the current gene
  plot_lin3 <- plotSmoothers(dat_sce, assays(dat_sce)$counts, gene = gene, 
                        curvesCols = curvesCols, border = FALSE) +
    ggplot2::scale_color_manual(values = curvesCols) +
    ggtitle(paste("Lineage 3 - ", gene)) +
    theme(legend.position = "none")  # Remove the legend
  
  # Add the plot to the list
  plots_list_lin3[[gene]] <- plot_lin3
}

# Create a PDF file to save all the plots
pdf("Lineage3_LYP_TOP50genesHSC1_youngvsold.pdf", width = 8, height = 10)  # Adjust width and height as needed

# Arrange and print all plots into the PDF
#plot_grid(plotlist = plots_list, nrow = 5, ncol = 3)  # Adjust the number of columns as needed

marrangeGrob(grobs=plots_list_lin3, nrow=5, ncol=3)
# Close the PDF device
dev.off()
```

###TOP50GENES LINEAGE4
```{r}
# Define colors for curves (assuming the same colors for all genes)
curvesCols <- c( "transparent", "transparent", "transparent", "transparent", "transparent", "transparent","transparent", "transparent", "transparent", "#0000FFFF","green", "#FF0000FF")

# Initialize an empty list to store individual plots
plots_list_lin4 <- list()

# Loop through each gene and create a plot
for (gene in Top50_HSC1genes_youngvsold) {
  # Create the plot for the current gene
  plot_lin4 <- plotSmoothers(dat_sce, assays(dat_sce)$counts, gene = gene, 
                        curvesCols = curvesCols, border = FALSE) +
    ggplot2::scale_color_manual(values = curvesCols) +
    ggtitle(paste("Lineage 4 - ", gene)) +
    theme(legend.position = "none")  # Remove the legend
  
  # Add the plot to the list
  plots_list_lin4[[gene]] <- plot_lin4
}

# Create a PDF file to save all the plots
pdf("Lineage4_MKP_TOP50genesHSC1_youngvsold.pdf", width = 8, height = 10)  # Adjust width and height as needed

# Arrange and print all plots into the PDF
#plot_grid(plotlist = plots_list, nrow = 5, ncol = 3)  # Adjust the number of columns as needed

marrangeGrob(grobs=plots_list_lin4, nrow=5, ncol=3)
# Close the PDF device
dev.off()
```
